Cambios Reto 5.1

+++ src/main/java/com/rentcloud/cloud/app/services/report/CountClient.java
@@ -13,10 +13,10 @@ import com.rentcloud.cloud.app.entities.Client;
  */
 public class CountClient {
     
-    private Double total;
+    private Long total;
     private Client client;
 
-    public CountClient(Double total, Client client) {
+    public CountClient(Long total, Client client) {
         this.total = total;
         this.client = client;
     }
@@ -25,11 +25,11 @@ public class CountClient {
  
     }
 
-    public Double getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Double total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
